made the cache reactive
